fix safari annotation export - closes #162
When clicking on export annotations as csv (or json) Safari
12.0.1 opens a new tab and shows an error message. This
patch removes the target attribute from the html a tag. Without
the target attribute the download attribute is honoured and
the file is saved. I also used the click method which works.
Tested on Safari 12.0.1 and Chrome 70.0.3538.102.

This patch closes issue #162:
https://gitlab.com/vgg/via/issues/162