fixed wrong region location for high zoom levels for large images (Closes: #166)
When using high zoom levels with large images and you draw
a region at the end of the image, then the region location
is not computed correctly. The reason is the wrong scaling of
the canvas which is computed from the initial image size which
is rounded based on initial scale computation at image load.

Closes: https://gitlab.com/vgg/via/issues/166